Revert "Merge branch 'main' into adding-colors-to-filter"
This reverts commit 7520c6f83140c8f3f5ff8a436a32a27cc497646e, reversing
changes made to 1b38e83864a534467294c4229a6b11fbeebd3a7b.
