+++ static/form.js
@@ -194,15 +194,12 @@ function submitForm() {
   document.getElementById('destAirport').remove();
   document.querySelector('label[for="startAirport"]').remove();
 
-  for (let i = 2; i <= 4; i++) {
-      document.getElementById('destinationLocation' + i);
-  }
   document.getElementById('submitButton').remove();
   document.getElementById('addButton').remove();
-  document.getElementById('additionalDestinations').remove();
-    
+
   // Remove all the additional destinations
-  
+  document.getElementById('additionalDestinations').remove();
+
   var hiddenElements = document.querySelectorAll('.hiddenObj');
 
   // Display loading animation for the user.
@@ -434,10 +431,4 @@ document.getElementById('dataForm').addEventListener('submit', function(event) {
 
   checkInputs();
 });
-
-// Function to return to the form page.
-function returnToForm() {
-  window.location.href = "/";
-}
-
