+++ static/requirements.txt
@@ -13,5 +13,3 @@ pytz==2024.1
 requests==2.31.0
 urllib3==2.2.1
 Werkzeug==3.0.1
-google-ai-generativelanguage==0.6.1
-flask_session==0.8.0
