+++ templates/base.html
@@ -6,8 +6,7 @@
     <title>NOTAM Sorter</title>
     <link rel="stylesheet" href="{{ url_for('static', filename='bootstrap/css/bootstrap.css') }}">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-datepicker@1.9.0/dist/css/bootstrap-datepicker.min.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/jquery-datetimepicker/2.5.20/jquery.datetimepicker.min.css" integrity="sha512-..." crossorigin="anonymous" referrerpolicy="no-referrer">
-    
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/jquery-datetimepicker/2.5.20/jquery.datetimepicker.min.css" integrity="sha512-..." crossorigin="anonymous" referrerpolicy="no-referrer">
     <!-- Custom styles for our project. -->
     <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
 
@@ -19,15 +18,6 @@
 
      <!-- Library for icons -->
      <script src="https://kit.fontawesome.com/ba0f5ae90e.js" crossorigin="anonymous"></script>
-     
-    <!-- Display script file reference. -->
-    <script src="static/display.js" ></script>
-
-    <!-- Library for icons -->
-    <script src="https://kit.fontawesome.com/ba0f5ae90e.js" crossorigin="anonymous"></script>
-    
-    <!-- Form script file reference. -->
-    <script src="static/display.js" ></script>
 </head>
 <body>
